Widen the quiet history table to int32 to stop it overflowing
History is updated with += depth*depth, so the counter for a frequently
refuted quiet move grows quadratically with depth.  In an int16_t table
it wraps past 32767 into a large negative value and the best quiet move
sorts last - the opposite of what the heuristic intends.

Measured peak history value in a single search:

  depth 12      580   (bench depth - safe)
  depth 18   11,690   (safe)
  depth 22   60,078   OVERFLOWS int16_t

So this is reachable in ordinary long time controls, not just in theory.
At depth 22 on a quiet middlegame the fix cuts the tree measurably:

  int16   30,595,129 nodes  score cp 409
  int32   28,501,093 nodes  score cp 412   (-6.8% nodes)

Bench is depth 12 and unchanged (8380015 nodes); depth 18 is also
identical, since neither reaches the overflow.  Table grows to 32 KiB.

+++ src/history.h
@@ -4,7 +4,10 @@
 #include <cstdint>
 
 struct HistoryTables {
-    int16_t quiet[2][64][64] = {};
+    // int32 rather than int16: updates are += depth*depth, so at high depth a
+    // handful of cutoffs on the same move overflow a 16-bit counter and the
+    // best quiet move wraps negative, sorting it last.  32 KiB total.
+    int32_t quiet[2][64][64] = {};
 };
 
 #endif
